Экспорт из хедер
вынес в отдельную функцию рендер по клику на лого
