+++ src/js/loadStartGallery.js
@@ -6,13 +6,7 @@ import transformGenres from './transfomGenres';
 // номер страницы открываемы по умолчанию
 const page = 1;
 
-// нашли лого
-const homeLogo = document.querySelector('.logo__link');
-
 const galleryList = document.querySelector('.gallery-list');
-
-// слушаем нажатие на logo
-homeLogo.addEventListener('click', renderClickLogo, false);
 
 // функция рендера
 function loadStartGallery(data) {
@@ -26,18 +20,15 @@ function loadStartGallery(data) {
 // функция получения данных с сервера и коррекция даты и жанров
 function dataRequest(page) {
   API.getTrendingMovie(page).then(({ resultsTrending }) => {
+    // генерим дату
     transformData(resultsTrending);
+    // генерим жанры
     transformGenres(resultsTrending);
+    // рендерим на страницу
     loadStartGallery(resultsTrending);
   });
 }
 // вызов функции рендера галереи при загрузке страницы
 dataRequest(page);
 
-// функция рендера галереи без перезагрузки страницы
-function renderClickLogo(evt) {
-  evt.preventDefault();
-  dataRequest(1);
-}
-
-export { loadStartGallery };
+export { loadStartGallery, dataRequest };
